Added network frontend and fixed navbar/footer. Made small refactor to make isUserAuthenticated easier to use. Added network page to controller.
