+++ synapsis/src/main/java/com/soen/synapsis/appuser/AppUser.java
@@ -1,5 +1,9 @@
 package com.soen.synapsis.appuser;
 
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+
 import javax.persistence.*;
 
 @Entity
@@ -21,7 +25,8 @@ public class AppUser {
     @Enumerated(EnumType.STRING)
     private Role role;
 
-    protected AppUser() {}
+    protected AppUser() {
+    }
 
     public AppUser(Long id, String name, String password, String email, Role role) {
         this.id = id;
@@ -76,6 +81,15 @@ public class AppUser {
 
     public void setRole(Role role) {
         this.role = role;
+    }
+
+    public static boolean isUserAuthenticated() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        if (auth == null || auth instanceof AnonymousAuthenticationToken)
+            return false;
+
+        return true;
     }
 
     @Override
