+++ synapsis/src/main/java/com/soen/synapsis/appuser/connection/ConnectionController.java
@@ -2,12 +2,13 @@ package com.soen.synapsis.appuser.connection;
 
 import com.soen.synapsis.appuser.AppUser;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
 
-@RestController
+@Controller
 public class ConnectionController {
 
     private final ConnectionService connectionService;
@@ -17,13 +18,20 @@ public class ConnectionController {
         this.connectionService = connectionService;
     }
 
+    @GetMapping("/network")
+    public String viewNetwork() {
+        if (!AppUser.isUserAuthenticated()) {
+            return "redirect:/";
+        }
+        return "pages/network";
+    }
+
     @PostMapping("/makeConnection")
     public String makeConnection(@RequestBody AppUser appUser1, @RequestBody AppUser appUser2, Model model) {
         try {
             String returnString = connectionService.makeConnection(appUser1, appUser2);
             return returnString;
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             model.addAttribute("error", "There was an error connecting with the user: " + e.getMessage());
             return "pages/home";
         }
