+++ synapsis/src/main/resources/static/css/styles.css
@@ -57,6 +57,8 @@ body {
     color: var(--color-black);
     line-height: 1.7;
     overflow: scroll;
+    min-height: 100vh;
+    position: relative;
 }
 
 h1, h2 {
