+++ synapsis/src/main/resources/templates/_layout.html
@@ -1,8 +1,9 @@
 <!DOCTYPE html>
-<html xmlns:th="https://www.ultraq.net.nz/thymeleaf/layout">
+<html xmlns:th="https://www.ultraq.net.nz/thymeleaf/layout" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
 <head>
     <meta charset="UTF-8">
     <title>Synapsis</title>
+    <script src="https://kit.fontawesome.com/6ebd7b3ed7.js" crossorigin="anonymous"></script>
     <script src="https://cdn.tailwindcss.com"></script>
     <script>
         tailwind.config = {
