+++ synapsis/src/main/resources/templates/fragments/footer.html
@@ -1,5 +1,5 @@
 <div th:fragment="footer" class="h-64">
-    <div class="w-full h-64 bg-black-variant absolute left-0">
+    <div class="w-full h-64 bg-black-variant absolute bottom-0 left-0">
         <div class="container flex flex-row justify-between mt-8">
             <div class="flex-none mt-4 w-3/6">
                 <img src="/assets/synapsis_red.svg" class="h-12 w-48 mr-10" alt="Synapsis Logo">
