+++ synapsis/src/main/resources/templates/fragments/navbar.html
@@ -1,5 +1,6 @@
 <div th:fragment="navbar" class="h-32">
-    <nav class="bg-white px-4 lg:px-2 py-6 fixed w-full z-20 top-0 left-0">
+    <!--Not logged in navbar-->
+    <nav sec:authorize="!isAuthenticated()" class="bg-white px-4 lg:px-2 py-6 fixed w-full z-20 top-0 left-0">
         <div class="container flex flex-wrap items-center justify-between mx-auto">
             <a href="/" class="flex items-center">
                 <img src="/assets/synapsis_black.svg" class="h-12 lg:h-16 mt-3 mr-3 " alt="Synapsis Logo">
@@ -16,7 +17,7 @@
                            aria-current="page">Home</a>
                     </li>
                     <li>
-                        <a href="/network"
+                        <a href="#"
                         >About</a>
                     </li>
                     <li>
@@ -27,6 +28,55 @@
                         <a href="#"
                         >Contact</a>
                     </li>
+                </ul>
+            </div>
+        </div>
+    </nav>
+
+    <!--Logged in navbar-->
+    <nav sec:authorize="isAuthenticated()" class="bg-white px-4 lg:px-2 py-2 fixed w-full z-20 top-0 left-0">
+        <div class="container flex flex-wrap items-center justify-between mx-auto">
+            <a href="/" class="flex items-center">
+                <img src="/assets/synapsis_black.svg" class="h-12 lg:h-16 mt-3 mr-3 " alt="Synapsis Logo">
+            </a>
+            <div class="items-center justify-between hidden w-full md:flex md:w-auto md:order-1" id="navbar-sticky">
+                <ul class="flex flex-col justify-between p-4 mt-4 md:flex-row md:space-x-6 md:mt-0 text-lg font-light text-black w-[1000px]">
+                    <a href="#" aria-current="page">
+                        <li class="flex flex-col items-center">
+                            <i class="fa-solid fa-house fa-2x h-10"></i>
+                            Home
+                        </li>
+                    </a>
+                    <a href="/network">
+                        <li class="flex flex-col items-center">
+                            <i class="fa-solid fa-users fa-2x h-10"></i>
+                            Network
+                        </li>
+                    </a>
+                    <a href="#">
+                        <li class="flex flex-col items-center">
+                            <i class="fa-solid fa-briefcase fa-2x h-10"></i>
+                            Jobs
+                        </li>
+                    </a>
+                    <a href="#">
+                        <li class="flex flex-col items-center">
+                            <i class="fa-regular fa-comment-dots fa-2x h-10"></i>
+                            Messages
+                        </li>
+                    </a>
+                    <a href="#">
+                        <li class="flex flex-col items-center">
+                            <i class="fa-solid fa-bell fa-2x h-10"></i>
+                            Notifications
+                        </li>
+                    </a>
+                    <a href="#">
+                        <li class="flex flex-col items-center">
+                            <i class="fa-solid fa-user fa-2x h-10"></i>
+                            My Profile
+                        </li>
+                    </a>
                 </ul>
             </div>
         </div>
