+++ synapsis/src/main/resources/templates/pages/network.html
@@ -6,24 +6,24 @@
         <div class="col-start-2 col-end-4 row-start-1 row-end-4 bg-light-50 rounded-2xl">
             <div class="ml-4 mt-3 mr-4">
                 <div class="font-sans text-3xl font-light mb-4">Connections</div>
-                <div class="flex flex-row justify-between items-end mt-3">
-                    <img class="flex-none h-8 w-8" src="" alt="people">
+                <a href="#" class="flex flex-row justify-between items-end mt-3">
+                    <i class="flex-none h-4 w-8 fa-solid fa-users fa-lg"></i>
                     <div class="flex-initial font-sans text-xl font-light w-40 ml-2">All
                         Connections
                     </div>
                     <div class="flex-none font-sans text-xl font-light w-8 text-right">1</div>
-                </div>
-                <div class="flex flex-row justify-between items-center mt-3">
-                    <img class="flex-none h-8 w-8" src="" alt="mail">
+                </a>
+                <a href="#" class="flex flex-row justify-between items-center mt-3">
+                    <i class="flex-none h-1 w-8 fa-solid fa-envelope fa-lg"></i>
                     <div class="flex-initial font-sans text-xl font-light w-40 ml-2">Requests</div>
                     <div class="flex-none font-sans text-xl font-light w-8 text-right">3</div>
-                </div>
-                <div class="flex flex-row justify-between items-center mt-3">
-                    <img class="flex-none h-8 w-8" src="" alt="suggestions">
+                </a>
+                <a href="#" class="flex flex-row justify-between items-center mt-3">
+                    <i class="flex-none h-2 w-8 fa-solid fa-hand-holding fa-lg"></i>
                     <div class="flex-initial font-sans text-xl font-light w-40 ml-2">Suggestions
                     </div>
                     <div class="flex-none font-sans text-xl font-light w-8 text-right">99+</div>
-                </div>
+                </a>
             </div>
         </div>
         <div class="col-start-5 col-end-12 row-span-1 h-14 bg-white border rounded-md">
